Added RabbitMQ for async communication

+++ src/Services/Cart/CartAPI/Models/ShoppingCart.cs
@@ -9,7 +9,7 @@ namespace CartAPI.Models
 
         public List<ShoppingCartItem> Items { get; set; } = new();
 
-        public double TotalPrice => Items.Sum(x => x.Price*x.Quantity);
+        public decimal TotalPrice => Items.Sum(x => x.Price*x.Quantity);
 
         public ShoppingCart(string userName)
         {
